dev(narugo): refactor the renderings
